Support editing and storing centre notes
Changes how corner notes are displayed. They were arranged in a "ring"
around the cell, with each note being in a specific position (1 is
top left, 2 is top middle, etc).

That doesn't work for centre notes, since the middle row of the corner
notes (4, 5, and 6) would overlap any centre notes. Instead, corner
notes are placed in the first available slot in 2 rows of 3 columns in
each cell.

This allows for 6 corner notes and 6 centre notes in each cell.

General refactoring:

- Update references to "note" to "corner note" or similar everywhere.

Specific changes:

- Modify Cell to support setting and getting centre notes as well as
  corner notes.

- Draw centre notes when present.

- Add an EditCellCentreNoteCommand.

- Clear centre as well as corner notes when clearing all notes.

- Modify the database to store, read, and write centre notes.